# openspec/openapi-v1.yaml
openapi: 3.1.0
info:
  title: AstrBot OpenAPI v1
  version: 0.1.0
  description: >
    Target REST contract for migrating AstrBot HTTP APIs to /api/v1.
    Dynamic AstrBot configuration payloads are intentionally modeled as open
    JSON objects because their schemas are provided at runtime by template
    endpoints.
servers:
  - url: http://localhost:6185/api/v1
    description: Local AstrBot server
security:
  - ApiKeyAuth: []
...
paths:
...
  /system-config:
    get:
      tags: [System Config]
      summary: Get the system configuration
      operationId: getSystemConfig
      x-astrbot-scope: config
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    put:
      tags: [System Config]
      summary: Replace the system configuration
      operationId: updateSystemConfig
      x-astrbot-scope: config
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DynamicConfig"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
...
  /config-profiles:
    get:
      tags: [Config Profiles]
      summary: List configuration profiles
      operationId: listConfigProfiles
      x-astrbot-scope: config
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    post:
      tags: [Config Profiles]
      summary: Create a configuration profile
      operationId: createConfigProfile
      x-astrbot-scope: config
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConfigProfileRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /config-profiles/{config_id}:
    get:
      tags: [Config Profiles]
      summary: Get a configuration profile
      operationId: getConfigProfile
      x-astrbot-scope: config
      parameters:
        - $ref: "#/components/parameters/ConfigId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    put:
      tags: [Config Profiles]
      summary: Replace a configuration profile
      operationId: updateConfigProfileContent
      x-astrbot-scope: config
      parameters:
        - $ref: "#/components/parameters/ConfigId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DynamicConfig"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    patch:
      tags: [Config Profiles]
      summary: Rename a configuration profile
      operationId: renameConfigProfile
      x-astrbot-scope: config
      parameters:
        - $ref: "#/components/parameters/ConfigId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenameRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    delete:
      tags: [Config Profiles]
      summary: Delete a configuration profile
      operationId: deleteConfigProfile
      x-astrbot-scope: config
      parameters:
        - $ref: "#/components/parameters/ConfigId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /config-routes:
    get:
      tags: [Bot Config Routes]
      summary: List UMO to configuration profile bindings
      operationId: listConfigRoutes
      x-astrbot-scope: config
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    put:
      tags: [Bot Config Routes]
      summary: Replace all UMO to configuration profile bindings
      operationId: replaceConfigRoutes
      x-astrbot-scope: config
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfigRoutesReplaceRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /config-routes/{umo}:
    put:
      tags: [Bot Config Routes]
      summary: Bind a UMO to a configuration profile
      operationId: upsertConfigRoute
      x-astrbot-scope: config
      parameters:
        - $ref: "#/components/parameters/Umo"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfigRouteUpsertRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    delete:
      tags: [Bot Config Routes]
      summary: Remove a UMO configuration profile binding
      operationId: deleteConfigRoute
      x-astrbot-scope: config
      parameters:
        - $ref: "#/components/parameters/Umo"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Ok:
      description: Standard AstrBot success response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SuccessEnvelope"
...
  parameters:
...
    ConfigId:
      name: config_id
      in: path
      required: true
      schema:
        type: string
...
    Umo:
      name: umo
      in: path
      required: true
      description: URL-encoded unified message origin.
      schema:
        type: string
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          const: ok
        message:
          type: string
        data: {}
      additionalProperties: false
...
    DynamicConfig:
      type: object
      additionalProperties: true
...
    CreateConfigProfileRequest:
      type: object
      properties:
        name:
          type: string
        config:
          $ref: "#/components/schemas/DynamicConfig"
      additionalProperties: false
    RenameRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
      additionalProperties: false
...
    ConfigRoutesReplaceRequest:
      type: object
      required: [routing]
      properties:
        routing:
          type: object
          additionalProperties:
            type: string
      additionalProperties: false
    ConfigRouteUpsertRequest:
      type: object
      required: [config_id]
      properties:
        config_id:
          type: string
          description: Use "default" to remove a custom route and fall back.
      additionalProperties: false
...